python3Packages.jaxlib: 0.3.22 -> 0.4.4
As part of this upgrade the builds for aarch64-linux and x86_64-darwin have been dropped. aarch64-linux was broken to begin with. x86_64-darwin may be recoverable but I do not have access to an x86_64-darwin machine in order to build and test changes. However, x86_64-darwin users can still use jaxlib through the jaxlib-bin package which maintains support for all platforms for which JAX releases jaxlib wheels.